data_processing/visualization/requirements.txt
pyqtgraph>=0.13.0
PyOpenGL>=3.1.6
pyserial>=3.5
numpy>=1.21.0
PyQt5>=5.15.0